requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.6
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
